refactor: rename env to config
